requirements.txt
Flask==3.0.3
ping3==4.0.8
python-dotenv==1.0.1